vote/requirements.txt
Flask
Redis
gunicorn
azure-storage==0.32.0